move data source v2 to catalyst module
