+++ sql/catalyst/src/main/java/org/apache/spark/sql/vectorized/ColumnarMap.java
